% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_date.R
\name{change_date}
\alias{change_date}
\alias{change_day}
\alias{change_month}
\alias{change_year}
\title{Change \code{Date} components}
\usage{
change_date(x, date)

change_day(x, day)

change_month(x, month)

change_year(x, year)
}
\arguments{
\item{x}{A \code{\link[base:Dates]{Date}}, \code{\link[base:DateTimeClasses]{POSIXct}}, or
\code{\link[base:DateTimeClasses]{POSIXlt}} object.}

\item{date}{A \code{\link[base:character]{character}} or \code{\link[base:Dates]{Date}}
object representing the new date.}

\item{day}{A \code{\link[base:numeric]{numeric}} object representing the day of
the month.}

\item{month}{A \code{\link[base:numeric]{numeric}} object representing the month
of the year.}

\item{year}{A \code{\link[base:numeric]{numeric}} object representing the year.}
}
\value{
A \code{\link[base:Dates]{Date}}, \code{\link[base:DateTimeClasses]{POSIXct}}, or
\code{\link[base:DateTimeClasses]{POSIXlt}} object with the new day, month, year,
or date.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}

\verb{change_*()} allows you to change the day, month, year, or date of a
\code{\link[base:Dates]{Date}}, \code{\link[base:DateTimeClasses]{POSIXct}}, or
\code{\link[base:DateTimeClasses]{POSIXlt}} object.
}
\examples{
## Scalar example

change_day(as.Date("2021-01-01"), 15)
#> [1] "2021-01-15" # Expected

change_month(as.POSIXct("2021-01-01 12:00:00", tz = "UTC"), 12)
#> [1] "2021-12-01 12:00:00 UTC" # Expected

change_year(as.POSIXlt("2021-01-01 12:00:00", tz = "UTC"), 2022)
#> [1] "2022-01-01 12:00:00 UTC" # Expected

change_date(
  x = as.POSIXlt("2021-01-01 12:00:00", tz = "UTC"),
  date = as.Date("2000-01-01")
)
#> [1] "2000-01-01 12:00:00 UTC" # Expected

## Vector example

x <- as.Date(c("2021-01-01", "2021-01-01"))
change_day(x, 15)
#> [1] "2021-01-15" "2021-01-15" # Expected

x <- as.POSIXct(c("2021-01-01 12:00:00", "2021-01-01 12:00:00"), tz = "UTC")
change_month(x, 10:11)
#> [1] "2021-10-01 12:00:00 UTC" "2021-11-01 12:00:00 UTC" # Expected

x <- as.POSIXlt(c("2021-01-01 12:00:00", "2021-01-01 12:00:00"), tz = "UTC")
change_year(x, 2022)
#> [1] "2022-01-01 12:00:00 UTC" "2022-01-01 12:00:00 UTC" # Expected

x <- as.POSIXlt(c("2021-01-01 12:00:00", "2021-01-01 12:00:00"), tz = "UTC")
change_date(x, as.Date("2000-01-01"))
#> [1] "2000-01-01 12:00:00 UTC" "2000-01-01 12:00:00 UTC" # Expected
}
\seealso{
Other utility functions: 
\code{\link{extract_seconds}()},
\code{\link{fix_hms}()},
\code{\link{flat_posixt_date}()},
\code{\link{get_last_week}()},
\code{\link{round_time}()}
}
\concept{utility functions}
