% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_unit.R
\name{convert_to_unit}
\alias{convert_to_unit}
\title{Convert a numerical object to a specific unit}
\usage{
convert_to_unit(x, from = "seconds", to = "hours")
}
\arguments{
\item{x}{A \code{\link[base:numeric]{numeric}} object.}

\item{from}{A string indicating the unit of \code{x} (see
the \emph{Units} section to learn more) (default: \code{"seconds"}).}

\item{to}{(optional) a string indicating the desire output unit
(see the \emph{Units} section to learn more) (default: \code{"hours"}).}
}
\value{
A \code{\link[base:numeric]{numeric}} object representing the decimals
hours \code{x}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{convert_to_unit()} allows you to convert a numerical object to different
kind of units.

See the \emph{Units} section for more information about the units that
\code{convert_to_unit()} can handle.
}
\section{Units}{


Here are the units that \code{convert_to_unit()} can handle for now:
\itemize{
\item years.
\item weeks.
\item days.
\item hours.
\item minutes.
\item seconds.
\item milliseconds.
\item microseconds.
\item nanoseconds.
\item degrees.
\item radians.
}
}

\examples{
convert_to_unit(60, "seconds", "minutes")
#> [1] 1

convert_to_unit(1, "minute", "hours")
#> [1] 0.01666667

convert_to_unit(1, "hour", "days")
#> [1] 0.0416666

convert_to_unit(1, "day", "weeks")
#> [1] 0.1428571

convert_to_unit(7, "days", "weeks")
#> [1] 1

convert_to_unit(1, "week", "days")
#> [1] 7
}
\concept{parsing/conversion functions}
