% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_hms.R
\name{fix_hms}
\alias{fix_hms}
\title{Transform \code{24:00:00} to \code{00:00:00} for \code{hms} objects}
\usage{
fix_hms(x)
}
\arguments{
\item{x}{A \code{\link[hms:hms]{hms}} vector.}
}
\value{
A \code{\link[hms:hms]{hms}} vector.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}

\code{fix_hms()} is a quick fix for \code{\link[hms:hms]{hms}} objects that have a
time of \code{24:00:00}. This function transforms the time to \code{00:00:00} for
consistency with the \code{\link[base:DateTimeClasses]{POSIXt}} class.
}
\examples{
hms::parse_hm("24:00") |> fix_hms()
#> 00:00:00 # Expected

hms::parse_hm(c("02:00", "24:00", "18:00")) |> fix_hms()
02:00:00 # Expected
00:00:00
18:00:00
}
\seealso{
Other utility functions: 
\code{\link{change_date}()},
\code{\link{extract_seconds}()},
\code{\link{flat_posixt_date}()},
\code{\link{get_last_week}()},
\code{\link{round_time}()}
}
\concept{utility functions}
