% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flat_posixt.R
\name{flat_posixt_date}
\alias{flat_posixt_date}
\alias{flat_posixt_hour}
\title{Adjust the date or time of a \code{POSIXt} vector to a specific date or hour}
\usage{
flat_posixt_date(x, base = as.Date("1970-01-01"), force_tz = TRUE, tz = "UTC")

flat_posixt_hour(x, base = hms::parse_hm("00:00"), force_tz = TRUE, tz = "UTC")
}
\arguments{
\item{x}{A \code{\link[base:DateTimeClasses]{posixt}} vector.}

\item{base}{A \code{\link[base:Dates]{Date}} (for \code{flat_posixt_date()} or
\code{\link[hms:hms]{hms}} (for \code{flat_posixt_hour()}) value (Default:
\code{as.Date("1970-01-01")}
(\href{https://en.wikipedia.org/wiki/Unix_time}{Unix Epoch})
or \code{hms::parse_hm("00:00")}).}

\item{force_tz}{A \code{\link[base:logical]{logical}} flag indicating whether to
force the timezone of the output to \code{tz} (Default: \code{TRUE}).}

\item{tz}{A \code{\link[base:character]{character}} string indicating the timezone
to use when \code{force_tz} is \code{TRUE} (Default: \code{"UTC"}).}
}
\value{
A \code{\link[base:DateTimeClasses]{posixt}} vector.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}

\verb{flat_posixt_*} functions adjust the date or time of a
\code{\link[base:DateTimeClasses]{POSIXt}} vector to a specific date or hour, with
\code{UTC} as the default timezone.

These functions are particularly useful for performing time arithmetic on
time objects that lack a date reference (e.g., \code{\link[hms:hms]{hms}} objects).
}
\examples{
as.POSIXct("2020-01-01 05:55:55", tz = "America/Sao_Paulo") |>
  flat_posixt_date()
#> [1] "1970-01-01 05:55:55 UTC" # Expected

c(
  as.POSIXct("2020-01-01 05:55:55", tz = "America/Sao_Paulo"),
  as.POSIXct("2020-01-01 18:40:05", tz = "America/Sao_Paulo")
) |>
 flat_posixt_date()
#> [1] "1970-01-01 05:55:55 UTC" "1970-01-01 18:40:05 UTC" # Expected

as.POSIXct("2020-01-01 05:55:55", tz = "America/Sao_Paulo") |>
 flat_posixt_hour(base = hms::parse_hm("00:01"))
#> [1] "2020-01-01 00:01:00 UTC" # Expected

c(
 as.POSIXct("2020-01-01 05:55:55", tz = "America/Sao_Paulo"),
 as.POSIXct("2020-01-01 18:40:05", tz = "America/Sao_Paulo")
) |>
 flat_posixt_hour(base = hms::parse_hm("00:01"))
#> [1] "2020-01-01 00:01:00 UTC" "2020-01-01 00:01:00 UTC" # Expected
}
\seealso{
Other utility functions: 
\code{\link{change_date}()},
\code{\link{extract_seconds}()},
\code{\link{fix_hms}()},
\code{\link{get_last_week}()},
\code{\link{round_time}()}
}
\concept{utility functions}
