% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_last.R
\name{get_last_week}
\alias{get_last_week}
\alias{get_last_month}
\title{Get last week or month}
\usage{
get_last_week(date = Sys.Date())

get_last_month(date = Sys.Date())
}
\arguments{
\item{date}{(optional) A \code{\link[base:as.Date]{Date}} object with the reference
date (default: \code{\link[=Sys.Date]{Sys.Date()}}).}
}
\value{
An \code{\link[lubridate:interval]{Interval}} object with the last week or
month.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}

\code{get_last_week()} and \code{get_last_month()} return the last week or month
interval, respectively, based on a reference date.
}
\examples{
## Scalar example

get_last_week(as.Date("2024-12-06"))
#> [1] 2024-11-24 UTC--2024-11-30 UTC # Expected

get_last_month(as.Date("2024-12-06"))
#> [1] 2024-11-01 UTC--2024-11-30 UTC # Expected

## Vector example

get_last_week(as.Date(c("2024-12-06", "2024-12-15")))
#> [1] 2024-11-24 UTC--2024-11-30 UTC # Expected
#' [1] 2024-12-08 UTC--2024-12-14 UTC # Expected

get_last_month(as.Date(c("2024-12-06", "2024-11-15")))
#> [1] 2024-11-01 UTC--2024-11-30 UTC  # Expected
#> [2] 2024-10-01 UTC--2024-10-31 UTC # Expected
}
\seealso{
Other utility functions: 
\code{\link{change_date}()},
\code{\link{fix_hms}()},
\code{\link{flat_posixt_date}()},
\code{\link{round_time}()}
}
\concept{utility functions}
