% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_duration.R
\name{int_duration}
\alias{int_duration}
\title{Compute the duration of an \code{Interval} object}
\usage{
int_duration(int = NULL)
}
\arguments{
\item{int}{An \code{\link[lubridate:interval]{Interval}} vector.}
}
\value{
A \code{\link[lubridate:duration]{Duration}} vector.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}

\code{int_duration()} computes the duration of an
\code{\link[lubridate:interval]{Interval}} object
}
\examples{
lubridate::interval(
  lubridate::ymd_hms("2023-01-01 22:00:01", tz = "UTC"),
  lubridate::ymd_hms("2023-01-02 02:00:01", tz = "UTC"),
  tzone = "UTC"
) |>
  int_duration()
#> [1] "14400s (~4 hours)" # Expected

lubridate::interval(
  lubridate::ymd_hms("2023-01-01 22:00:01", tz = "UTC"),
  lubridate::ymd_hms("2023-01-02 02:00:01", tz = "UTC"),
  tzone = "UTC"
) |>
  int_duration() |>
  lubridate::as.period()
#> [1] "4H 0M 0S" # Expected

c(
  lubridate::interval(
    lubridate::ymd_hms("2023-01-01 22:00:01", tz = "UTC"),
    lubridate::ymd_hms("2023-01-02 02:00:01", tz = "UTC"),
    tzone = "UTC"
  ),
  lubridate::interval(
    lubridate::ymd_hms("1990-01-01 06:00:00", tz = "UTC"),
    lubridate::ymd_hms("1990-01-01 12:00:00", tz = "UTC"),
    tzone = "UTC"
  )
) |>
  int_duration()
#> [1] "14400s (~4 hours)" "21600s (~6 hours)" # Expected
}
\seealso{
Other Interval functions: 
\code{\link{int_mean}()},
\code{\link{int_overlap}()}
}
\concept{Interval functions}
