% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_mean.R
\name{int_mean}
\alias{int_mean}
\title{Compute the mean of an \code{Interval} object}
\usage{
int_mean(int)
}
\arguments{
\item{int}{An \code{\link[lubridate:interval]{Interval}} vector.}
}
\value{
An \code{\link[lubridate:interval]{Interval}} vector.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}

\code{int_mean()} computes the mean of an \code{\link[lubridate:interval]{Interval}}
object.
}
\examples{
lubridate::interval(
  lubridate::ymd_hms("2023-01-01 22:00:01", tz = "UTC"),
  lubridate::ymd_hms("2023-01-02 02:00:01", tz = "UTC"),
  tzone = "UTC"
) |>
  int_mean()
#> [1] "2023-01-02 00:00:01 UTC" # Expected

c(
  lubridate::interval(
    lubridate::ymd_hms("2023-01-01 22:00:01", tz = "UTC"),
    lubridate::ymd_hms("2023-01-02 02:00:01", tz = "UTC"),
    tzone = "UTC"
  ),
  lubridate::interval(
    lubridate::ymd_hms("1990-01-01 06:00:00", tz = "UTC"),
    lubridate::ymd_hms("1990-01-01 12:00:00", tz = "UTC"),
    tzone = "UTC"
  )
) |>
  int_mean()
#> [1] "2023-01-02 00:00:01 UTC" "1990-01-01 09:00:00 UTC" # Expected
}
\seealso{
Other Interval functions: 
\code{\link{int_duration}()}
}
\concept{Interval functions}
