% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_overlap.R
\name{int_overlap}
\alias{int_overlap}
\title{Compute the overlapping interval between two \code{Interval} objects}
\usage{
int_overlap(int_1, int_2)
}
\arguments{
\item{int_1, int_2}{\code{\link[lubridate:interval]{Interval}} vectors. If
they have a length of `1``, they will be recycled to the length of the
longest.}
}
\value{
An \code{\link[lubridate:interval]{Interval}} vector with the overlapping
period between \code{int_1} and \code{int_2}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}

\code{int_overlap()} computes the overlapping interval between two
\code{\link[lubridate:interval]{Interval}} objects.

See the \emph{Details} section for more information.
}
\details{
The overlapping interval is the time span where both intervals intersect.
If the intervals do not overlap, the function returns an \code{NA-NA} \code{Interval}.
The illustration below demonstrates the concept:

\if{html}{\out{<div class="sourceCode">}}\preformatted{          +______________________+_________+
          |                      |         |
          |           Interval 1 |         |
          +______________________+_________+
          |//////////////////////|
          |//// Overlap Area ////|
  +_______+//////////////////////+
  |       |                      |
  |       |  Interval 2          |
  +_______+______________________+
}\if{html}{\out{</div>}}
}
\examples{
int_1 <- lubridate::interval(
  start = lubridate::as_datetime("2020-01-01 00:00:00"),
  end = lubridate::as_datetime("2021-01-01 00:00:00"),
)

int_2 <- lubridate::interval(
  start = lubridate::as_datetime("2020-05-01 00:00:00"),
  end = lubridate::as_datetime("2021-05-01 00:00:00"),
)

int_3 <- lubridate::interval(
  start = lubridate::as_datetime("2021-01-01 00:00:00"),
  end = lubridate::as_datetime("2025-01-01 00:00:00"),
)

int_overlap(int_1, int_2)
#> [1] 2020-05-01 UTC--2021-01-01 UTC # Expected

int_overlap(int_1, c(int_2, int_3))
#> [1] 2020-05-01 UTC--2021-01-01 UTC # Expected
#> [2] 2021-01-01 UTC--2021-01-01 UTC
}
\seealso{
Other Interval functions: 
\code{\link{int_duration}()},
\code{\link{int_mean}()}
}
\concept{Interval functions}
