% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorter_interval.R
\name{longer_duration}
\alias{longer_duration}
\title{Find the shorter duration between two hours}
\description{
See \code{\link[mctq:shorter_interval]{shorter_duration()}} for details.

See \code{\link[mctq:shorter_interval]{longer_duration()}} for details.
}
\keyword{internal}
