% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorter_interval.R
\name{longer_interval}
\alias{longer_interval}
\title{Find the longer interval between two hours}
\description{
See \code{\link[mctq:shorter_interval]{longer_interval()}} for details.
}
\keyword{internal}
