% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lubridate-package.R
\docType{package}
\name{lubritime-package}
\alias{lubritime}
\alias{lubritime-package}
\title{lubritime: An Extension for lubridate}
\description{
An extension for the 'lubridate' package from tidyverse, adding new features to deal with temporal objects.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://danielvartan.github.io/lubritime/}
  \item \url{https://github.com/danielvartan/lubritime/}
  \item Report bugs at \url{https://github.com/danielvartan/lubritime/issues/}
}

}
\author{
\strong{Maintainer}: Daniel Vartanian \email{danvartan@gmail.com} (\href{https://orcid.org/0000-0001-7782-759X}{ORCID}) [conceptor, copyright holder]

}
\keyword{internal}
