% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_data.R
\name{raw_data}
\alias{raw_data}
\title{Get paths to \code{lubritime} raw data}
\usage{
raw_data(file = NULL)
}
\arguments{
\item{file}{(optional) a \code{character} object indicating the raw data file
name(s). If \code{NULL}, all raw data file names will be printed (default:
\code{NULL}).}
}
\value{
\itemize{
\item If \code{file = NULL}, a \code{character} object with all file names available.
\item If \code{file != NULL}, a string with the file name path.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{raw_data()} returns the raw data paths of the \code{lubritime} package.
}
\examples{
\dontrun{
## To list all raw data file names available

raw_data()

## To get the file path from a specific raw data

raw_data(raw_data()[1])}
}
\concept{utility functions}
