% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_duration.R
\name{int_duration}
\alias{int_duration}
\title{Compute the duration of an interval}
\usage{
int_duration(start, end)
}
\arguments{
\item{start}{A \code{\link[base:as.POSIXlt]{POSIXct}} object representing the start
of the interval.}

\item{end}{A \code{\link[base:as.POSIXlt]{POSIXct}} object representing the end of
the interval.}
}
\value{
A \code{\link[lubridate:period]{Period}} object representing the duration of
the interval.
}
\description{
This function computes the duration of an interval given its start and end
points.
}
\examples{
## Scalar example

int_duration(
  start = lubridate::dmy_hms("15/01/2023 09:57:35", tz = "UTC"),
  end = lubridate::dmy_hms("18/01/2023 23:59:35", tz = "UTC")
)
#> [1] "3d 14h 2m 0s" # Expected

## Vector example

int_duration(
  start = lubridate::dmy_hms(
    c("15/01/2023 09:57:35", "18/01/2023 09:57:35"),
    tz = "UTC"
  ),
  end = lubridate::dmy_hms(
    c("18/01/2023 23:59:35", "20/01/2023 23:59:35"),
    tz = "UTC"
  )
)
#> [1] "3d 14h 2m 0s" "2d 14h 2m 0s" # Expected
}
\concept{Interval functions}
