% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lubridate-package.R
\docType{package}
\name{lubritime-package}
\alias{lubritime}
\alias{lubritime-package}
\title{lubritime: An Extension for the Lubridate Package}
\description{
An extension for the 'lubridate' package from tidyverse, adding new features to deal with temporal objects.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gipso.github.io/lubritime/}
  \item \url{https://github.com/gipso/lubritime/}
  \item Report bugs at \url{https://github.com/gipso/lubritime/issues/}
}

}
\author{
\strong{Maintainer}: Daniel Vartanian \email{danvartan@gmail.com} (\href{https://orcid.org/0000-0001-7782-759X}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Mario Pedrazzoli \email{pedrazzo@usp.br} (\href{https://orcid.org/0000-0002-5257-591X}{ORCID}) [scientific advisor]
}

Other contributors:
\itemize{
  \item Interdisciplinary Sleep Research Group (GIPSO) [funder]
  \item University of Sao Paulo (USP) [funder]
}

}
\keyword{internal}
